Struct gazetta_core::model::Entry

pub struct Entry<EntryMeta> where
    EntryMeta: Meta,  {
    pub title: String,
    pub description: Option<String>,
    pub date: Option<Date>,
    pub index: Option<Index>,
    pub cc: Vec<String>,
    pub meta: EntryMeta,
    pub name: String,
    pub content: String,
    pub format: String,
}

An entry in the website.

Note: Pages do not correspond one-to-one to Entries (due to pagination).

Fields
title: String

The entry's title.

This is separate from the general metadata for sorting and linking. All entries should have titles.

description: Option<String>

The entry's description.

If present, this will be included in compact indices.

date: Option<Date>

The entry's date.

This is separate from the general metadata for sorting. Many entries will have dates.

index: Option<Index>

The entries index options (if specified).

cc: Vec<String>

Indices into which this entry should be linked.

meta: EntryMeta

Extra metadata.

name: String

The entry name.

content: String

The content

format: String

Content format

Methods

impl<EntryMeta> Entry<EntryMeta> where
EntryMeta: Meta,

pub fn from_file(full_path: PathBuf, name: &str) -> Result<Self, SourceError>