Struct gazetta_core::model::Source

pub struct Source<SourceMeta = (), EntryMeta = ()> where SourceMeta: Meta, EntryMeta: Meta {
    pub title: String,
    pub root: PathBuf,
    pub origin: String,
    pub prefix: String,
    pub entries: Vec<Entry<EntryMeta>>,
    pub static_entries: Vec<StaticEntry>,
    pub stylesheets: Vec<PathBuf>,
    pub javascript: Vec<PathBuf>,
    pub icon: Option<PathBuf>,
    pub meta: SourceMeta,
}

The Source object reads and interprets a source directory.

The fields are intentionally public. Feel free to manually generate or modify this structure.

Fields

title

The website's title.

By default, this field is read from gazetta.yaml.

root

The source root directory.

This is specified on construction.

origin

The website origin (http://mydomain.com:1234)

By default, this field is derived from the value of base in gazetta.yaml.

prefix

The directory under the origin at which this site will be hosted (e.g. "/").

By default, this field is derived from the value of base in gazetta.yaml.

entries

The website content to be rendered.

By default, this list is populated with Entries generated from files with the basename index under the root directory excluding:

Files under directories named "static".

Files under assets/.

static_entries

The website content to be deployed as-is (no rendering).

By default, this list is populated with directories under the root directory named "static" excluding:

Directories under directories named "static".

Directories under assets/.

stylesheets

The website stylesheets. When rendered, these will be concatinated into a single stylesheet.

By default, this list is populated by the files in is assets/stylesheets/ in lexicographical order.

javascript

The website javascript. When rendered, these will be concatenated into a single javascript file.

By default, this list is populated by the files in is assets/javascript/ in lexicographical order.

icon

The path website's icon.

By default, this points to assets/icon.png (if it exists).

meta

Additional metadata read from gazetta.yaml.

Methods

impl<SourceMeta, EntryMeta> Source<SourceMeta, EntryMeta> where SourceMeta: Meta, EntryMeta: Meta

fn build_index(&self, entry: &Entry<EntryMeta>) -> Vec<&Entry<EntryMeta>>

Build an index for an entry.

This index includes all entries that "cc" this entry and all entries specified in this entry's index pattern.

fn new<P: AsRef<Path>>(root: P) -> Result<Self, AnnotatedError<SourceError>>

Parse a source directory to create a new source.

fn reload(&mut self) -> Result<(), AnnotatedError<SourceError>>

Reload from the source directory.

Call this after changing source files.