Struct gazetta_core::view::Site

pub struct Site<'a, G> where G: Gazetta + 'a, G::SiteMeta: 'a {
    pub title: &'a str,
    pub origin: &'a str,
    pub prefix: &'a str,
    pub stylesheets: Option<&'a str>,
    pub javascript: Option<&'a str>,
    pub icon: Option<&'a str>,
    pub meta: &'a G::SiteMeta,
}

A "website"

You should include this view at the top of your websites "head". It renders into script, stylesheet, icon tags, metadata tags, and importantly the base tag.

html! {
    html {
        head {
            : site;
            // ...
        }
    }
}

Fields

title

The website's title

origin

The "canonical" origin for the website. (i.e. the <protocol>://<domain>:<port> part of the url)

prefix

The path prefix at wich we're serving this website.

stylesheets

The concatinated stylesheets.

javascript

The concatinated javascript.

icon

The icon.

meta

Extra metadata specified in the Source.